Switch to unified RustSystemAllocator for both tests and proofs
